:lipstick: update: refactor DetailMemberPage and social icon widget; improve layout and structure
